+++ CHANGELOG.md
@@ -7,16 +7,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [0.23.2] - 2021-06-19
+### Fixed
+- Relative url resolution [#105].
+
+## [0.23.1] - 2021-06-17
+### Fixed
+- Error using `search.pages()` without arguments.
+
+## [0.23.0] - 2021-06-17
 ### Added
 - New `slugify_urls` plugin.
+- Support for quotes to `page.search()`.
+  This allows to insert spaces in the values.
+  For example: `page.search("'tag with spaces' title='Title with spaces'")`.
+
+### Changed
+- BREAKING: The URLs of the pages no longer will be slugified by default.
+  Use `slugify_urls` plugin.
+- The plugin `attributes` is disabled by default.
 
 ### Removed
 - The `slugifyUrls` option. Use the `slugify_urls` plugin for that.
-- Support of objects in the `url` variable.
+- Support for objects in the `url` variable.
 
 ### Fixed
-- Updated `std`.
+- Replaced the `terser` dependency with a Deno version.
+- Updated `std` and `postcss`.
 
 ## [0.22.6] - 2021-06-15
 ### Added
@@ -34,8 +51,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 - Bug in the function to merge default and user options.
-- Updated `postcss` and `eta`.
 - Ensure the `Page` and `Directory` classes always have the `src` object.
+- Updated `postcss` and `eta`.
 
 ## [0.22.5] - 2021-06-10
 ### Added
@@ -864,8 +881,11 @@ The first version.
 [#95]: https://github.com/lumeland/lume/issues/95
 [#97]: https://github.com/lumeland/lume/issues/97
 [#98]: https://github.com/lumeland/lume/issues/98
+[#105]: https://github.com/lumeland/lume/issues/105
 
-[Unreleased]: https://github.com/lumeland/lume/compare/v0.22.6...HEAD
+[0.23.2]: https://github.com/lumeland/lume/compare/v0.23.1...v0.23.2
+[0.23.1]: https://github.com/lumeland/lume/compare/v0.23.0...v0.23.1
+[0.23.0]: https://github.com/lumeland/lume/compare/v0.22.6...v0.23.0
 [0.22.6]: https://github.com/lumeland/lume/compare/v0.22.5...v0.22.6
 [0.22.5]: https://github.com/lumeland/lume/compare/v0.22.4...v0.22.5
 [0.22.4]: https://github.com/lumeland/lume/compare/v0.22.3...v0.22.4